requirements.txt
Django==1.4
Fabric==1.4.3
PIL==1.1.7
boto==2.5.2
dj-database-url==0.2.1
-e git+https://github.com/sehmaschine/django-grappelli.git@d69c02e5ab6227040aef8d50f31b8c9c3513807c#egg=django_grappelli-dev
django-storages==1.1.4
django-tastypie==0.9.11
mimeparse==0.1.3
newrelic==1.3.0.289
psycopg2==2.4.5
pycrypto==2.6
pysolr==2.1.0-beta
python-dateutil==1.5
python-memcached==1.48
ssh==1.7.14
wsgiref==0.1.2
